refactor(deck): replace hover preview with click-to-modal in table views
Extract CardImageModal to a shared component, replace the desktop hover
image preview with click-to-open-modal on card names in both deck and
archetype table views. Add a subtle hover underline hint on card names
(.card-name-link). Remove the card-hover.ts dependency from the
archetype variant table view.
