+++ assets/components/DeckCardList.tsx
@@ -16,21 +16,18 @@
  * @see docs/features.md F2.23 — Interactive card mosaic with responsive grid
  */
 
-import React, { useCallback, useEffect, useMemo, useRef, useState } from 'react';
+import React, { useCallback, useMemo, useRef, useState } from 'react';
 import {
     ActionIcon,
-    CloseButton,
     Group,
     Loader,
-    Modal,
     SegmentedControl,
     Table,
     Text,
     Tooltip,
-    UnstyledButton,
 } from '@mantine/core';
-import { useMediaQuery } from '@mantine/hooks';
-import { IconCopy, IconCheck, IconShare, IconShoppingCart, IconChevronLeft, IconChevronRight } from '@tabler/icons-react';
+import { IconCopy, IconCheck, IconShare, IconShoppingCart } from '@tabler/icons-react';
+import CardImageModal, { type FlatCard } from './CardImageModal';
 import CardMosaicGrid from './CardMosaicGrid';
 
 interface CardData {
@@ -82,175 +79,21 @@ type Variant = 'original' | 'minified';
 type ViewMode = 'table' | 'mosaic';
 
 /**
- * Card image gallery modal — mobile swipeable viewer with prev/next navigation.
- *
- * @see docs/features.md F6.2 — TCGdex card data enrichment
+ * Clickable card name — opens the card image modal on click.
  */
-const CardImageModal: React.FC<{
-    opened: boolean;
-    cards: CardEntry[];
-    currentIndex: number;
-    onClose: () => void;
-    onNavigate: (index: number) => void;
-}> = ({ opened, cards, currentIndex, onClose, onNavigate }) => {
-    const touchStartX = useRef(0);
-    const touchStartY = useRef(0);
-
-    const card = cards[currentIndex];
-
-    const navigate = useCallback((direction: number) => {
-        onNavigate((currentIndex + direction + cards.length) % cards.length);
-    }, [currentIndex, cards.length, onNavigate]);
-
-    useEffect(() => {
-        if (!opened) return;
-
-        const handleKeyDown = (event: KeyboardEvent): void => {
-            if (event.key === 'ArrowLeft') navigate(-1);
-            if (event.key === 'ArrowRight') navigate(1);
-        };
-
-        document.addEventListener('keydown', handleKeyDown);
-
-        return () => document.removeEventListener('keydown', handleKeyDown);
-    }, [opened, navigate]);
-
-    if (!card) return null;
-
-    const title = `${card.quantity} \u00d7 ${card.name}`;
-
-    return (
-        <Modal
-            opened={opened}
-            onClose={onClose}
-            withCloseButton={false}
-            centered
-            size="sm"
-            padding={0}
-            styles={{
-                body: { padding: 0 },
-                content: { overflow: 'hidden' },
-            }}
-        >
-            <div style={{ display: 'flex', alignItems: 'center', justifyContent: 'space-between', padding: '8px 12px' }}>
-                <Text size="sm" fw={500} style={{ flex: 1 }}>{title}</Text>
-                <Text size="xs" c="dimmed" style={{ marginRight: 8 }}>
-                    {currentIndex + 1} / {cards.length}
-                </Text>
-                <CloseButton onClick={onClose} />
-            </div>
-            <div
-                style={{ position: 'relative', textAlign: 'center', padding: '0 8px 8px' }}
-                onTouchStart={(event) => {
-                    touchStartX.current = event.touches[0].clientX;
-                    touchStartY.current = event.touches[0].clientY;
-                }}
-                onTouchMove={(event) => {
-                    const deltaX = Math.abs(event.touches[0].clientX - touchStartX.current);
-                    const deltaY = Math.abs(event.touches[0].clientY - touchStartY.current);
-                    if (deltaY > deltaX) {
-                        event.preventDefault();
-                    }
-                }}
-                onTouchEnd={(event) => {
-                    const deltaX = event.changedTouches[0].clientX - touchStartX.current;
-                    if (Math.abs(deltaX) > 50) {
-                        navigate(deltaX < 0 ? 1 : -1);
-                    }
-                }}
-            >
-                <UnstyledButton
-                    onClick={() => navigate(-1)}
-                    style={{ position: 'absolute', left: 4, top: '50%', transform: 'translateY(-50%)', zIndex: 1 }}
-                >
-                    <IconChevronLeft size={28} />
-                </UnstyledButton>
-                <img
-                    src={card.imageUrl}
-                    alt={card.name}
-                    style={{ maxWidth: '100%', borderRadius: 4 }}
-                />
-                <UnstyledButton
-                    onClick={() => navigate(1)}
-                    style={{ position: 'absolute', right: 4, top: '50%', transform: 'translateY(-50%)', zIndex: 1 }}
-                >
-                    <IconChevronRight size={28} />
-                </UnstyledButton>
-            </div>
-        </Modal>
-    );
-};
-
-interface CardEntry {
-    imageUrl: string;
-    name: string;
-    quantity: number;
-}
-
-/**
- * Card hover preview — shows card image on mouse hover (desktop) or tap (mobile modal).
- */
-const CardName: React.FC<{ card: CardData; onMobileTap?: () => void }> = ({ card, onMobileTap }) => {
-    const [showPreview, setShowPreview] = useState(false);
-    const spanRef = useRef<HTMLSpanElement>(null);
-    const imgRef = useRef<HTMLImageElement>(null);
-    const isMobile = useMediaQuery('(max-width: 767px)');
-
-    const handleMouseEnter = useCallback(() => {
-        if (!spanRef.current || !imgRef.current) return;
-
-        const rect = spanRef.current.getBoundingClientRect();
-        const imgHeight = Math.min(Math.max(280, window.innerHeight / 3), 672);
-        const imgWidth = imgHeight / 1.4;
-
-        let top: number;
-        if (rect.top >= imgHeight + 8) {
-            top = rect.top - imgHeight;
-        } else {
-            top = rect.bottom + 4;
-        }
-
-        top = Math.max(4, Math.min(top, window.innerHeight - imgHeight - 4));
-
-        let left = rect.left;
-        if (left + imgWidth > window.innerWidth - 4) {
-            left = window.innerWidth - imgWidth - 4;
-        }
-        left = Math.max(4, left);
-
-        // Set position imperatively before showing to avoid flicker
-        imgRef.current.style.top = `${top}px`;
-        imgRef.current.style.left = `${left}px`;
-
-        setShowPreview(true);
-    }, []);
-
-    if (!card.imageUrl) {
+const CardName: React.FC<{ card: CardData; onClick?: () => void }> = ({ card, onClick }) => {
+    if (!card.imageUrl || !onClick) {
         return <>{card.cardName}</>;
     }
 
     return (
-        <span
-            ref={spanRef}
-            className="card-hover"
-            data-quantity={card.quantity}
-            onMouseEnter={handleMouseEnter}
-            onMouseLeave={() => setShowPreview(false)}
-            onClick={(event) => {
-                if (isMobile && onMobileTap) {
-                    event.preventDefault();
-                    onMobileTap();
-                }
-            }}
-        >
+        <span className="card-name-link" role="button" tabIndex={0} onClick={onClick} onKeyDown={(event) => {
+            if (event.key === 'Enter' || event.key === ' ') {
+                event.preventDefault();
+                onClick();
+            }
+        }}>
             {card.cardName}
-            <img
-                ref={imgRef}
-                className="card-hover-img"
-                src={card.imageUrl}
-                alt={card.cardName}
-                style={{ display: showPreview ? 'block' : undefined }}
-            />
         </span>
     );
 };
@@ -262,8 +105,8 @@ const CardSection: React.FC<{
     label: string;
     cards: CardData[];
     tableLabels: { qty: string; card: string; set: string };
-    onCardTap?: (card: CardData) => void;
-}> = ({ label, cards, tableLabels, onCardTap }) => (
+    onCardClick?: (card: CardData) => void;
+}> = ({ label, cards, tableLabels, onCardClick }) => (
     <div className="card shadow-sm">
         <div className="card-header card-header-themed">
             <h5 className="mb-0">{label}</h5>
@@ -283,7 +126,7 @@ const CardSection: React.FC<{
                         <Table.Tr key={`${card.cardName}-${card.setCode}-${card.cardNumber}-${index}`}>
                             <Table.Td>{card.quantity}</Table.Td>
                             <Table.Td>
-                                <CardName card={card} onMobileTap={onCardTap ? () => onCardTap(card) : undefined} />
+                                <CardName card={card} onClick={onCardClick ? () => onCardClick(card) : undefined} />
                             </Table.Td>
                             <Table.Td><code>{card.setCode}</code></Table.Td>
                             <Table.Td>{card.cardNumber}</Table.Td>
@@ -301,8 +144,8 @@ const CardSection: React.FC<{
 const CardTable: React.FC<{
     groupedCards: Record<string, CardData[]>;
     labels: Labels;
-    onCardTap?: (card: CardData) => void;
-}> = ({ groupedCards, labels, onCardTap }) => {
+    onCardClick?: (card: CardData) => void;
+}> = ({ groupedCards, labels, onCardClick }) => {
     const pokemonCards = groupedCards.pokemon ?? [];
     const trainerCards = groupedCards.trainer ?? [];
     const energyCards = groupedCards.energy ?? [];
@@ -314,21 +157,21 @@ const CardTable: React.FC<{
         <div className="row">
             <div className="col-md-6 mb-3">
                 {pokemonCards.length > 0 && (
-                    <CardSection label={labels.sectionPokemon} cards={pokemonCards} tableLabels={tableLabels} onCardTap={onCardTap} />
+                    <CardSection label={labels.sectionPokemon} cards={pokemonCards} tableLabels={tableLabels} onCardClick={onCardClick} />
                 )}
                 {energyInLeft && energyCards.length > 0 && (
                     <div className="mt-3">
-                        <CardSection label={labels.sectionEnergy} cards={energyCards} tableLabels={tableLabels} onCardTap={onCardTap} />
+                        <CardSection label={labels.sectionEnergy} cards={energyCards} tableLabels={tableLabels} onCardClick={onCardClick} />
                     </div>
                 )}
             </div>
             <div className="col-md-6 mb-3">
                 {trainerCards.length > 0 && (
-                    <CardSection label={labels.sectionTrainer} cards={trainerCards} tableLabels={tableLabels} onCardTap={onCardTap} />
+                    <CardSection label={labels.sectionTrainer} cards={trainerCards} tableLabels={tableLabels} onCardClick={onCardClick} />
                 )}
                 {!energyInLeft && energyCards.length > 0 && (
                     <div className="mt-3">
-                        <CardSection label={labels.sectionEnergy} cards={energyCards} tableLabels={tableLabels} onCardTap={onCardTap} />
+                        <CardSection label={labels.sectionEnergy} cards={energyCards} tableLabels={tableLabels} onCardClick={onCardClick} />
                     </div>
                 )}
             </div>
@@ -394,14 +237,14 @@ export const DeckCardList: React.FC<DeckCardListProps> = ({
         ? minifiedList
         : rawList;
 
-    const flatCards: CardEntry[] = useMemo(() => {
-        const entries: CardEntry[] = [];
+    const flatCards: FlatCard[] = useMemo(() => {
+        const entries: FlatCard[] = [];
         const order = ['pokemon', 'trainer', 'energy'];
 
         for (const section of order) {
             for (const card of activeCards[section] ?? []) {
                 if (card.imageUrl) {
-                    entries.push({ imageUrl: card.imageUrl, name: card.cardName, quantity: card.quantity });
+                    entries.push({ imageUrl: card.imageUrl, cardName: card.cardName, quantity: card.quantity });
                 }
             }
         }
@@ -409,8 +252,8 @@ export const DeckCardList: React.FC<DeckCardListProps> = ({
         return entries;
     }, [activeCards]);
 
-    const handleCardTap = useCallback((card: CardData) => {
-        const index = flatCards.findIndex((entry) => entry.imageUrl === card.imageUrl && entry.name === card.cardName);
+    const handleCardClick = useCallback((card: CardData) => {
+        const index = flatCards.findIndex((entry) => entry.imageUrl === card.imageUrl && entry.cardName === card.cardName);
         if (index >= 0) {
             setCardModalIndex(index);
             setCardModalOpen(true);
@@ -565,7 +408,7 @@ export const DeckCardList: React.FC<DeckCardListProps> = ({
 
             {/* Table view */}
             {!isMinifiedPending && viewMode === 'table' && (
-                <CardTable groupedCards={activeCards} labels={labels} onCardTap={handleCardTap} />
+                <CardTable groupedCards={activeCards} labels={labels} onCardClick={handleCardClick} />
             )}
 
             {/* Interactive mosaic grid view */}
